Share the enrollment-info view between munki and turbo
The device-facing enrollment-info endpoint (get the enrollment by its
secret in the Authorization header, return it, post the info-request
event) was duplicated between munki and turbo — the only two consumers.
Extract BaseEnrollmentInfoView next to its EnrollmentSecretAuthentication
partner in the inventory app; both modules' views now just declare their
auth class, serializer and queryset. Behavior-preserving (the event type
is derived from the view's own auth class).
